requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-allauth==0.57.0
django-environ==0.11.2
channels==4.0.0
channels-redis==4.1.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
openai==1.100.2
django-htmx==1.17.0
uvicorn==0.24.0
daphne==4.0.0
django-storages[google]==1.14.2
gunicorn==21.2.0